Remove unused NavMenu and appriopriately rename NavBar
